refactor cors implementation
