+++ src/main/java/gov/usgs/wma/waterdata/springinit/SpringConfig.java
@@ -1,0 +1,17 @@
+package gov.usgs.wma.waterdata.springinit;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class SpringConfig  implements WebMvcConfigurer {
+	@Override
+	public void addCorsMappings(CorsRegistry registry) {
+		registry
+				.addMapping("/**")
+				.allowedOrigins("*")
+				.allowedMethods("GET", "OPTIONS")
+				.allowedHeaders("Origin", "Accept", "X-Requested-With", "Content-Type", "Access-Control-Request-Method", "Access-Control-Request-Headers");
+	}
+}
